# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Profile API
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users:
    # binds a127 app logic to a route
    x-swagger-router-controller: basic
    get:
      description: Returns all profiles
      # used as the method name of the controller
      operationId: findAll
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Users"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"    
    post:
      description: Add a profile
      operationId: insertOne
      parameters: 
        - in: body
          name: body
          description: inserted profile
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{id}/profile:
    x-swagger-router-controller: basic
    get:
      description: get a profile
      operationId: findOne
      parameters:
        - in: path
          name: id
          description: get by id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/User"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    put:
      description: update a profile
      operationId: updateOne
      parameters:
        - in: path
          name: id
          description: filter
          required: true
          type: string
        - in: body
          name: body
          description: updated
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/User"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: delete a profile
      operationId: deleteOne
      parameters:
        - in: path
          name: id
          description: filter
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/User"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{pid}/profile/educations:
    x-swagger-router-controller: education
    get:
      description: get all education data form a user
      operationId: findAllEducation
      parameters: 
        - name: pid
          description: user's pid
          in: path
          required: true
          type: string
      responses: 
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/Educations"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a education
      operationId: insertOneEducation
      parameters: 
        - name: pid
          description: user's pid
          in: path
          required: true
          type: string
        - name: body
          in: body
          schema:
            $ref: "#/definitions/Education"
      responses:
        "200":
          description: Success
          schema: 
            $ref: "#/definitions/Education"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{pid}/profile/educations/{id}:
    x-swagger-router-controller: education
    get:
      description: find a education
      operationId: findOneEducation
      parameters:
        - name: pid
          description: user's pid
          in: path
          type: string
          required: true
        - name: id
          description: education's id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: Sucess
          schema:
            $ref: "#/definitions/Education"
        default: 
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a education
      operationId: updateOneEducation
      parameters: 
        - name: pid
          description: user's id
          in: path
          required: true
          type: string
        - name: id
          description: education's id
          in: path
          required: true
          type: string
        - name: body
          description: education for update
          in: body
          schema: 
            $ref: "#/definitions/Education"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Education"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: remove a education
      operationId: removeOneEducation
      parameters:
        - name: pid
          description: user's pid
          in: path
          required: true
          type: string
        - name: id
          description: education's id
          in: path
          required: true
          type: string
      responses:
        "200": 
          description: Success
          schema: 
            $ref: "#/definitions/Education"
        default: 
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  User:
    type: object
    required:
      - user_name
      - create_date
      - update_date
    properties:
      user_name:
        type: string
      avatar_file_id:
        type: string
      avatar_coordinate:
        type: string
      cover_file_id:
        type: string
      cover_coordinate:
        type: string
      complete_status:
        type: string
      introduction:
        type: string
      personal_url:
        type: string
      create_date:
        type: string
      update_date:
        type: string
  Users:
    type: array
    items: 
      $ref: "#/definitions/User"
  Education:
    type: object
    required:
      - pid
      - school_name
      - major_name
      - degree
      - edu_location
      - start_year
      - start_month
      - create_date
      - update_date
    properties:
      pid:
        type: string
      school_name:
        type: string
      major_name:
        type: string
      degree:
        type: string
      edu_location:
        type: string
      start_year:
        type: string
      start_month:
        type: string
      end_year:
        type: string
      end_month:
        type: string
      status:
        type: string
      create_date:
        type: string
      update_date:
        type: string
  Educations:
    type: array
    items:
      $ref: "#/definitions/Education"